refactor: 페이지 폴더 생성 및 App 코드 Payments 컴포넌트로 이동
